Rama con ejercicio 3.8

+++ src/main/java/org/iesvdm/dao/ComercialDAOImpl.java
@@ -41,7 +41,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 			ps.setString(idx++, comercial.getNombre());
 			ps.setString(idx++, comercial.getApellido1());
 			ps.setString(idx++, comercial.getApellido2());
-			ps.setDouble(idx++, comercial.getComision());
+			ps.setDouble(idx++, comercial.getComision().doubleValue());
 			return ps;
 		},keyHolder);
 
@@ -61,7 +61,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                 							  rs.getString("nombre"), 
                 							  rs.getString("apellido1"),
                 							  rs.getString("apellido2"), 
-                							  rs.getFloat("comisión"))
+                							  rs.getBigDecimal("comisión"))
                 						 	
         );
 		
@@ -78,7 +78,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 								rs.getString("nombre"),
 								rs.getString("apellido1"),
 								rs.getString("apellido2"),
-								rs.getFloat("comisión"))
+								rs.getBigDecimal("comisión"))
 						, id
 				);
